+fix IntegerMath.power +FloatMath.log(a,b);

+++ jfixby-cmns-api/jfixby-commons-api/com/jfixby/cmns/api/math/IntegerMath.java
@@ -68,7 +68,7 @@ public class IntegerMath {
 		return invoke().index(b);
 	}
 
-	public static long power(long a, long n) {
+	public static long power(long a, int n) {
 		return invoke().power(a, n);
 	}
